# fix: improve high log throughput job speed