--- a/pkg/shell/process.go
+++ b/pkg/shell/process.go
@@ -348,12 +348,12 @@ func (p *Process) writeCommandToFile(cmdFilePath, command string) error {
 
 // Determines the buffer size used when reading command output from the TTY.
 // Note: the TTY read does not block until this buffer is completely filled,
-// so there's no need to use a small buffer here. Using a 2-8k one seems reasonable.
+// so there's no need to use a small buffer here. 1-4k proved to be a good choice.
 func (p *Process) readBufferSize() int {
 	rand.Seed(time.Now().UnixNano())
 
-	min := 2048
-	max := 8192
+	min := 1024
+	max := 4096
 
 	// #nosec
 	return rand.Intn(max-min) + min