--- a/pkg/shell/process_test.go
+++ b/pkg/shell/process_test.go
@@ -39,6 +39,22 @@ func Benchmark__CommandOutput_100K(b *testing.B) {
 	}
 }
 
+func Benchmark__CommandOutput_250K(b *testing.B) {
+	p := createProcess(b, fmt.Sprintf("for i in {0..500}; do echo '%s'; done", strings.Repeat("x", 1024)))
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		p.Run()
+	}
+}
+
+func Benchmark__CommandOutput_500K(b *testing.B) {
+	p := createProcess(b, fmt.Sprintf("for i in {0..500}; do echo '%s'; done", strings.Repeat("x", 1024)))
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		p.Run()
+	}
+}
+
 func Benchmark__CommandOutput_1M(b *testing.B) {
 	p := createProcess(b, fmt.Sprintf("for i in {0..1000}; do echo '%s'; done", strings.Repeat("x", 1024)))
 	b.ResetTimer()